integration of frame position publisher for hand gaze following
The frame position publisher is stopped by default and needs to be started
via the button or switch in the task panel. Frame names can also be changed
in the task panel. These frame positions are published and received by the
"oh-gaze-tracking-demo", which needs to be started in a separate terminal.